+++ InventoryTools/Compendium/Windows/CompendiumListWindow.cs
@@ -190,6 +190,11 @@ public class CompendiumListWindow : CompendiumWindow
                                 }
                             }
 
+                            if (ImGui.MenuItem("Generate Support Dump"))
+                            {
+                                this.MediatorService.Publish(new OpenGenericWindowMessage(typeof(SupportDumpWindow)));
+                            }
+
                             if (ImGui.MenuItem("Report a Issue"))
                             {
                                 "https://github.com/Critical-Impact/InventoryTools".OpenBrowser();