+++ InventoryTools/Services/GameCraftSources/IGameCraftSource.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace InventoryTools.Services.GameCraftSources
+{
+    public interface IGameCraftSource
+    {
+        IReadOnlyList<GameCraftCategory> GetAvailableCategories();
+    }
+}