+++ InventoryTools/Ui/Windows/FiltersWindow.cs
@@ -577,6 +577,11 @@ namespace InventoryTools.Ui
                                 }
                             }
 
+                            if (ImGui.MenuItem("Generate Support Dump"))
+                            {
+                                this.MediatorService.Publish(new OpenGenericWindowMessage(typeof(SupportDumpWindow)));
+                            }
+
                             if (ImGui.MenuItem("Report a Issue"))
                             {
                                 "https://github.com/Critical-Impact/InventoryTools".OpenBrowser();