+++ InventoryToolsMock/MockStartup.cs
@@ -1,5 +1,8 @@
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using AllaganLib.GameSheets.Sheets;
+using CriticalCommonLib.Crafting;
 using InventoryTools.Services;
 using Microsoft.Extensions.Hosting;
 
@@ -13,7 +16,7 @@ public class MockStartup : IHostedService
     {
         this.windowService = windowService;
     }
-    
+
     public Task StartAsync(CancellationToken cancellationToken)
     {
         windowService.OpenWindow<MockWindow>();