+++ InventoryTools/CHANGELOG.md
@@ -8,6 +8,15 @@ Instead the changelog reader and automation surrounding plugin PRs will add the 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.html).
 
+## [15.0.9] - 2026-06-14
+
+### Added
+- Added "Support Dump" feature, making it easier for plugin users to provide their logs + config
+- Added a new "Add From Game" menu to the Crafts list replacing the "Add Company Crafts" button. This currently supports the Company Craft parts menu and the Grand Company Supply/Provision windows.
+
+### Changed
+- Lumina Supplemental data update
+
 ## [15.0.8] - 2026-06-14
 
 ### Added